http://tools.yalamanchili.info:8080/browse/SSTO-1579 update leave request support
git-svn-id: http://tools.sstech.us/svn/office/trunk@4344 04aae24c-3eb2-d842-8c26-b4285eb26770

+++ office-master/office-bpm/src/main/java/info/yalamanchili/office/bpm/OfficeBPMTaskService.java
@@ -7,6 +7,7 @@
  */
 package info.yalamanchili.office.bpm;
 
+import com.google.common.base.Strings;
 import info.chili.security.dao.CRoleDao;
 import info.chili.service.jrs.types.Entry;
 import info.chili.spring.SpringContext;
@@ -114,6 +115,9 @@ public class OfficeBPMTaskService {
     }
 
     public void deleteAllTasksForProcessId(String processId, boolean deleteProcessInstance) {
+        if (Strings.isNullOrEmpty(processId)) {
+            return;
+        }
         for (Task task : getTasksForProcessId(processId)) {
             bpmTaskService.deleteTask(task.getId());
         }
